models.py:
Added the Post model to save all posts created. The model still requires
some changes to able to store images.

views.py:

Added a View class for the home page and a ListView class for displaying
all blog posts created by me.

templates:

Added a _header.html and _footer.html file and included it in the base.html
file to create the navbar and footer of the web app. Created a home.html
file (home page of the website) and posts.html file to show all posts.